Start charts from minimum value
Per information visualization best practices (I'll look up a source if
needed) the minimum possible value should be at the bottom of the chart
if the maximum is at the top of the chart.

Since the questions receive 1 to 5 points each, the minimum value for
all charts on the site is max/5. Make it the default for BarChart and
BarChartCompare, and add the `min` prop for explicit configuration.